fix(runtime): make CPU model imports Triton-optional
